requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
convopyro @ git+https://github.com/Ripeey/Conversation-Pyrogram@a39cc9e0c1a689c9904dd5be5419246868daecfb
Faker==24.4.0
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-RESTful==0.3.10
Flask-WTF==1.2.1
frozenlist==1.4.1
g4f==0.2.6.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
outcome==1.3.0.post0
pyaes==1.6.1
pycparser==2.21
PyJWT==2.8.0
pyrocon==2.0.6
Pyrogram==2.0.106
pyshorteners==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.29
TgCrypto==1.2.5
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
undetected-chromedriver==3.5.5
urllib3==2.2.1
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.4
